Disable the `unix` module on Android.
The only methods it define require libc::SO_REUSEPORT,
which current versions on the libc crate does not provide for Android targets.

This is a temporary work around to make this crate build at all on Android.
It should be reverted
when https://github.com/rust-lang-nursery/libc/pull/70 lands.

This not a breaking change since the crate did not build at all before.