Prevent keepalive_ms and set_keepalive_ms from always failing on OpenBSD.
OpenBSD (as of release 6.6) does not support the socket option TCP_KEEPIDLE.
To be using SO_KEEPALIVE was not correct and was failing at runtime.

NetBSD supports the option and it should therefore be treated like generic
unix.  The use of SO_KEEPALIVE as the KEEPALIVE_OPTION on NetBSD also
does not seem to be correct.